backend/requirements.txt
fastapi
uvicorn
pydantic
motor
python-dotenv 